requirements.txt
pandas
numpy
PyYAML
scikit-learn
scipy
matplotlib
seaborn
joblib
requests
python-dotenv
azure-storage-blob
